pyproject.toml
[project]
name = "budget"
version = "0.1.0"
description = "Personal Budget Tracker - A gRPC-based application for tracking daily transactions, account balances, spending limits, and financial analytics"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy>=2.0.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
budget-grpc-server = "budget.grpc_server:serve"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--cov=budget",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
